# assistant/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Assistant backend service API",
        "title": "Assistant API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:4321",
    "basePath": "/",
    "paths": {
...
        "/api/llm/v1/chat/completions": {
            "post": {
                "tags": [
                    "LLM代理"
                ],
                "summary": "聊天补全(自动路由到可用供应商)",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/llmproxy.proxyChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/llm/v1/models/{model}": {
            "get": {
                "tags": [
                    "LLM代理"
                ],
                "summary": "查询模型详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型ID",
                        "name": "model",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "llmproxy.proxyChatRequest": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {}
                },
                "model": {
                    "type": "string"
                },
                "stream": {
                    "type": "boolean"
                }
            }
        },
...
    }
}